+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ClinicalVariableCard/CategoricalVariableCard.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import {
   compose,
+  lifecycle,
   setDisplayName,
   withProps,
   withPropsOnChange,
@@ -349,4 +350,9 @@ export default compose(
       )),
     })
   ),
+  lifecycle({
+    componentDidMount() {
+      console.log(`${this.props.fieldName} componentDidMount`);
+    },
+  }),
 )(EnhancedClinicalVariableCard);
