+++ src/packages/@ncigdc/modern_components/ClinicalAnalysis/ContinuousTestComponent.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+class ContinuousTestComponent extends React.Component {
+  componentDidMount() {
+    console.log('hello world componentDidMount');
+  }
+
+  render() {
+    // console.log('ContinuousTestComponent');
+    return (
+      <div>Hello world!</div>
+    );
+  }
+}
+
+export default ContinuousTestComponent;
